+++ src/wp/ap/gtk/ap_UnixDialog_Goto.cpp
@@ -750,7 +750,7 @@ AP_UnixDialog_Goto::updateXMLIDList( GtkWidget* w )
         GtkTreeIter iter;
         std::set< std::string > xmlids;
         rdf->getAllIDs( xmlids );
-        UT_DEBUGMSG (("MIQ: xmlids.sz:%lu\n", xmlids.size() ));
+        UT_DEBUGMSG (("MIQ: xmlids.sz:%lu\n", (long unsigned)xmlids.size() ));
 
         for( std::set< std::string >::iterator xiter = xmlids.begin();
              xiter != xmlids.end(); ++xiter )
